+++ src/main/php/lang/ast/emit/ReadonlyClasses.class.php
@@ -33,9 +33,11 @@ trait ReadonlyClasses {
       }
 
       // Prevent dynamic members
-      $throw= new Code('throw new \\Error("Cannot create dynamic property ".__CLASS__."::".$name);');
       $context= $result->codegen->enter(new InType($class));
-      $context->virtual[null]= [$throw, $throw];
+      $context->virtual[null]= [
+        new Code('trigger_error("Undefined property: ".__CLASS__."::\$".$name, E_USER_WARNING); return $_;'),
+        new Code('throw new \\Error("Cannot create dynamic property ".__CLASS__."::".$name);')
+      ];
     }
 
     return parent::emitClass($result, $class);